Rename: Add subdirectory for register page
회원가입 페이지를 위한 서브디렉토리 생성 후 회원가입 페이지 파일 옮김
